datapath/config: Own node configuration notifier
Move the local node datapath configuration notifier out of node manager and
provide it from a dedicated datapath configuration cell.

Update the Linux datapath, loader, neighbor discovery, and integration tests
to consume the relocated notifier.

REVIEW: Check that pkg/datapath/config is the right long-term owner for this
notifier and its lifecycle restriction.

AIL:3
Related: https://github.com/cilium/cilium/issues/41744
