+++ pkg/datapath/loader/util_test.go
@@ -31,7 +31,6 @@ import (
 	"github.com/cilium/cilium/pkg/endpointstate"
 	"github.com/cilium/cilium/pkg/hive"
 	"github.com/cilium/cilium/pkg/maps/registry"
-	"github.com/cilium/cilium/pkg/node/manager"
 	"github.com/cilium/cilium/pkg/option"
 	"github.com/cilium/cilium/pkg/promise"
 )
@@ -87,7 +86,7 @@ func newTestLoader(tb testing.TB) *loader {
 		cell.Provide(func() (
 			sysctl.Sysctl,
 			linuxConfig.Writer,
-			*manager.NodeConfigNotifier,
+			*config.NodeConfigNotifier,
 			promise.Promise[endpointstate.Restorer],
 			prefilter.PreFilter,
 		) {
@@ -95,7 +94,7 @@ func newTestLoader(tb testing.TB) *loader {
 			resolver.Resolve(&FakeRestorer{})
 			return sysctl.NewDirectSysctl(afero.NewOsFs(), "/proc"),
 				configWriterForTest(tb),
-				&manager.NodeConfigNotifier{},
+				&config.NodeConfigNotifier{},
 				promise,
 				&FakePreFilter{}
 		}),
